Changed the placement of some files
